refactor: share the router log-capture test helper
WHY:
Every router log written as one plain string needs a test that reads
the formatter's own bytes, to prove no ANSI escape reached the payload.
The comparison log kept that subscriber setup inside its own test
module, so the next log to be tested would copy it.

HOW:
- move the capturing writer and the subscriber setup into a test-only
  log_capture module
- give it capture_payloads, which runs a closure and returns what
  followed a prefix on each line
- rewrite the comparison-log test helper to call it
